Move flock package to pkg/flock
