+++ pkg/flock/flock_solaris.go
